+++ database/migrations/2026_05_08_104238_create_reviews_table.php
@@ -6,23 +6,17 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
             $table->string('name');
             $table->text('review');
-            $table->integer('rating');
+            $table->unsignedTinyInteger('rating');
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('reviews');
